+++ scanapi/settings.py
@@ -7,7 +7,7 @@ DEFAULT_CONFIG_PATH = ".scanapi.yaml"
 
 class Settings(dict):
     def __init__(self):
-        self["spec_path"] = "Scanfile"
+        self["spec_path"] = "scanapi.yaml"
         self["output_path"] = None
         self["template"] = None